requirements.txt
pytube==10.5.1
requests==2.25.1
Pillow==8.1.0
spotipy==2.17.0
eyed3==0.9.6